backend/pyproject.toml
[project] 
name = "app" 
version = "0.1.0" 
description = "Enterprise Analytics Platform Backend" 
authors = [{ name = "Architect", email = "architect@example.com" }] 
requires-python = ">=3.12" 
dependencies = [ 
    "alembic>=1.13.1", 
    "asyncpg>=0.29.0", 
    "duckdb>=0.10.0", 
    "fastapi>=0.109.0", 
    "httpx>=0.26.0", 
    "mpax>=0.2.4", 
    "passlib[argon2]>=1.7.4", 
    "pydantic-settings>=2.1.0", 
    "pydantic[email]>=2.6.0", 
    "python-jose[cryptography]>=3.3.0", 
    "python-multipart>=0.0.9", 
    "sqlalchemy[asyncio]>=2.0.25", 
    "sqlglot>=28.6.0", 
    "uvicorn[standard]>=0.27.0", 
] 

[dependency-groups] 
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "respx>=0.22.0",
] 

[build-system] 
requires = ["hatchling"] 
build-backend = "hatchling.build" 

[tool.ruff] 
exclude = [ 
    ".eggs", 
    ".git", 
    ".ipynb_checkpoints", 
    ".pyenv", 
    ".pytest_cache", 
    ".pytype", 
    ".ruff_cache", 
    ".tox", 
    ".venv", 
    ".vscode", 
    "_build", 
    "build", 
    "dist", 
    "site-packages", 
    "venv", 
] 
line-length = 122
indent-width = 2

[tool.ruff.lint] 
select = ["E4", "E7", "E9", "F"] 
ignore = [] 
fixable = ["ALL"] 
unfixable = [] 

# Allow unused variables when underscore-prefixed. 
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$" 

[tool.ruff.format] 
quote-style = "double" 
indent-style = "space" 
skip-magic-trailing-comma = false
line-ending = "auto" 
docstring-code-format = false
docstring-code-line-length = "dynamic" 

[tool.hatch.build.targets.wheel] 
packages = ["src/app"] 

[tool.pytest.ini_options] 
pythonpath = ["src"] 
testpaths = ["tests"] 
markers = ["asyncio: mark test as async"] 
# Configure pytest-asyncio for auto-discovery and session-wide loop to support session fixtures
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
